requirements.txt
binance_futures_connector==4.0.0
hexbytes==0.3.1
hmx_v2_python==1.2.1
matplotlib==3.8.4
numpy==1.24.3
pandas==2.2.2
pybit==5.6.2
Pypubsub==4.0.3
python-dotenv==1.0.1
python_binance==1.0.17
Requests==2.32.3
setuptools==68.2.2
synthetix==0.1.10
web3==6.16.0
